requirements.txt
beautifulsoup4==4.9.3
chardet==4.0.0
idna==2.10
six==1.15.0
soupsieve==2.1
urllib3==1.26.2
python-craigslist-headless==0.0.5
